Web-Scraping/SeleniumScrape/requirements.txt
selenium>=4.23
webdriver-manager>=4.0
python-dotenv>=1.0